pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "ansys-scadeone"
dynamic = ["version"]
description = "Pythonic Ansys Scade One"
readme = "README.md"
requires-python = ">=3.7,<4"
license = {file = "LICENSE"}
authors = [
    {name = "ANSYS, Inc.", email = "pyansys.maintainers@ansys.com"}
]
maintainers = [
    {name = "Scade One developers", email = "xavier.fornari@ansys.com"}
#    {name = "ANSYS, Inc.", email = "pyansys.maintainers@ansys.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "pythonnet>=3.0.1",
    "typing>=3.7.4.3",
    "typing_extensions>=4.0.1",
    "platformdirs>=3.6.0"
]

[project.optional-dependencies]
tests = [
    "pytest>=7.1.0",
    "pytest-cov>=3.0.0",
]
doc = [
    "ansys-sphinx-theme==0.12.0",    
    "numpydoc>=1.2",
    "sphinx>=7.1.0",
    "sphinx-copybutton>=0.4",
    "sphinx-autobuild==2021.3.14",
    "sphinx-autodoc-typehints==1.23.0",
    "pytest-sphinx==0.5.0",
]
checks = [
    "codespell==2.2.6",
    "flake8==6.1.0"
]

[tool.flit.module]
name = "ansys.scadeone"

[project.urls]
Source = "https://github.com/ansys/pyscadeone"

[project.scripts]
pyscadeone = "ansys.scadeone.cli:main"

# Formatting
[tool.black]
line-length = 100

[tool.isort]
profile = "black"
force_sort_within_sections = true
line_length = 100
default_section = "THIRDPARTY"
src_paths = ["doc", "src", "tests"]

[tool.codespell]
skip = '*.pyc,*.txt,*.gif,*.png,*.jpg,*.js,*.html,*.doctree,*.ipynb,flycheck*,./.git/*,*.yml,doc/_build/*,doc/images/*,dist/*'
ignore-words = "spelling_ok.txt"
quiet-level = 3

# Test & Coverage
[tool.coverage.run]
source = ["pyscadeone"]

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
log_file = "logs/pytest-logs.txt"
